pyproject.toml
[tool.poetry]
name = "show-case-bot"
version = "0.1.0"
description = ""
authors = ["arsenyvolodko <arvolodko@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.21.0"
python-dotenv = "1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
